Done with refactor - Delete files not used
