--- src/constants/index.ts
@@ -1,1 +1,0 @@
-export const OUTFUNNEL_URL = 'https://sink.outfunnel.com';
